requirements.txt
Scrapy==2.6.1
python-dotenv==0.20.0
Telethon==1.24.0
